requirements.txt
stmpy
paho-mqtt
appJar
sounddevice
soundfile